AuthService/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic>=2
pydantic-settings>=2
redis>=5
pydantic[email]
bcrypt==4.0.1